requirements.txt
amqp==2.3.2
argon2-cffi==18.3.0
atomicwrites==1.2.1
attrs==18.2.0
beautifulsoup4==4.6.3
billiard==3.5.0.4
celery==4.2.1
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
defusedxml==0.5.0
Django==2.1.3
django-allauth==0.38.0
django-cors-headers==2.4.0
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
django-decouple==2.1
django-extensions==2.1.3
django-rest-framework==0.1.0
django-webpack-loader==0.6.0
djangorestframework==3.9.0
factory-boy==2.11.1
Faker==1.0.0
idna==2.7
inflection==0.3.1
kombu==4.2.1
lxml==4.2.5
more-itertools==4.3.0
multitasking==0.0.7
numpy==1.15.4
oauthlib==2.1.0
pandas==0.23.4
pandas-datareader==0.7.0
pluggy==0.8.0
psycopg2==2.7.6.1
py==1.7.0
pycparser==2.19
pytest==4.0.0
python-dateutil==2.7.5
python3-openid==3.1.0
pytz==2018.7
Quandl==3.4.4
requests==2.20.1
requests-oauthlib==1.0.0
six==1.11.0
sqlparse==0.2.4
text-unidecode==1.2
urllib3==1.24.1
vine==1.1.4
whitenoise==4.1.1
wrapt==1.10.11
